Create Wine Functionality Implemented.
Built out a form to CREATE a bottle of wine and automatically display
it in the handlebars output. Used a lot of materializeCSS functionality
here including sliders.
Reformatted a lot of JPEGS since some were loading too slow.

Next up is UPDATE an existing wine functionality and then the DELETE
functionality.