Completed functionality of Update Wine.
This code contains the first working version of an UPDATE working
successfully from the front end to the back end. Validated CURL script
first and then wrote the enclosed logic.
This still needs to be cleaned up from a UX perspective but the core
logic of doing an update to an existing bottle of wine works fine.

Next up - Delete wine.