Third and final attempt at relative path to images on GitHub.
Moved images folder inside the assets folder to see if this works.

If this doesn't work then I'll use the 'raw link' approach to images
on GitHub that appears in several issue logs. This must be some quirk
with how github refers to images stored in a folder in a deployed app.

https://github.com/ga-wdi-boston/full-stack-project/issues/787